Added hash and printing functions for BDDs
