+++ libraries/symbolic/include/mcrl2/symbolic/ldd_stream.h
@@ -12,6 +12,7 @@
 
 #ifdef MCRL2_ENABLE_SYLVAN
 
+#include "mcrl2/symbolic/utility.h"
 #include "mcrl2/utilities/indexed_set.h"
 #include "mcrl2/utilities/bitstream.h"
 
@@ -19,22 +20,6 @@
 
 #include <deque>
 #include <iostream>
-
-namespace std
-{
-
-/// \brief specialization of the standard std::hash function.
-template<>
-struct hash<sylvan::ldds::ldd>
-{
-  std::size_t operator()(const sylvan::ldds::ldd& ldd) const
-  {
-    std::hash<std::uint64_t> hasher;
-    return hasher(ldd.get());
-  }
-};
-
-} // namespace std
 
 namespace mcrl2::symbolic
 {
